Rename modules to lowercase according to PEP 8
